+++ ultraphotos/Views/VideoPlayerView.swift
@@ -1,0 +1,26 @@
+//
+//  VideoPlayerView.swift
+//  ultraphotos
+//
+//  Created by Friedrich Ewald on 3/14/26.
+//
+
+import SwiftUI
+import AVKit
+
+struct VideoPlayerView: NSViewRepresentable {
+    let player: AVPlayer
+
+    func makeNSView(context: Context) -> AVPlayerView {
+        let playerView = AVPlayerView()
+        playerView.controlsStyle = .inline
+        playerView.player = player
+        return playerView
+    }
+
+    func updateNSView(_ nsView: AVPlayerView, context: Context) {
+        if nsView.player !== player {
+            nsView.player = player
+        }
+    }
+}